Impl basic roles and authorization (with tests)
- Add roles and valid roles to user
- change cbSend in _api to always return json
- impl api endpoint to toggle users in roles
- confirm only admins can toggle roles
- change test sessions to use a session object instead of users
- create user clones for idempotent tests
- Impl auth middleware check
- add roles to session (users still need to logout and log back in)
- change session api to users api
- add unauthorized route
- impl getUsersInRole in user service
- init test db with an administrator user in the admin role
- change test login helper to take user object and add it into the
global list to pass back

+++ server/services/_service.js
@@ -36,13 +36,15 @@ export default function(model, logging) {
       })
     },
     update: (id, data, cb) => {
+      if (!id) return cb(new Error('Cannot update object by null id'), null) 
       var ups = _.omit(data, '_id') // so mongo doesn't complain
       model.findByIdAndUpdate(id, ups).lean().exec( (e, r) => {
-        if (e || logging) { $log('svc.update.error', e, data) }
+        if (e || logging) { $log('svc.update.error', id, e, data) }
         cb(e, r)
       })
     },
     deleteById: (id, cb) => {
+      if (!id) return cb(new Error('Cannot delete object by null id'), null) 
       model.findByIdAndRemove(id, (e) => {
         if (e || logging) { $log('svc.delete', e) }
         cb(e)
